Add groups extraction
Many SAML providers, such as ADFS and Auth0, put in information about
what groups the user is in, and typically in the claim called
'http://schemas.xmlsoap.org/claims/Group'.

This change adds the ability to specify an XPath to extract the groups
information, in a similar fashion to the username extraction. This is done
via a configuration parameter called `xpath_groups_location`, which has
as its default the value 'http://schemas.xmlsoap.org/claims/Group'.